refactor(p2p): split desired-state engine tests
